\name{regionGoodnessOfFit-methods}
\docType{methods}
\alias{regionGoodnessOfFit-methods}
\alias{regionGoodnessOfFit,data.frame-method}
\alias{regionGoodnessOfFit,ExpData-method}
\alias{regionGoodnessOfFit}

\title{ Calculate goodness-of-fit statistics }
\description{
  A generic method for calculating chi-squared goodness-of-fit statistics.  Dispatches on either a \code{data.frame}
  or and \code{expData} object.
}
\section{Methods}{
  \describe{
    \item{\code{
	signature(obj = "data.frame")
    }}{
    }
    \item{\code{
	signature(obj = "ExpData")}}{
    }
}}
\keyword{methods}

